[llm_d] postprocess: llm_d: move around the CSV dashboard field names and AnalysisConfig
