+++ projects/llm_d/postprocess/llm_d/plugin.py
@@ -8,6 +8,7 @@ from typing import Any
 
 import yaml
 
+from projects.caliper.engine.kpi.analyze import AnalysisConfig
 from projects.caliper.engine.model import (
     ParseResult,
     PostProcessingPlugin,
@@ -17,69 +18,27 @@ from projects.caliper.engine.model import (
 from projects.guidellm.postprocess.guidellm.dashboard import (
     deployment_metadata_from_profile,
     enrich_guidellm_parse_result,
-    export_dashboard_kpis_to_csv,
-    normalize_product_version,
 )
 from projects.guidellm.postprocess.guidellm.plugin import GuideLLMPlugin
-from projects.guidellm.postprocess.guidellm.plugin import analysis_config as analysis_config
 from projects.llm_d.orchestration.render_inference_service import _build_vllm_args
 from projects.llm_d.orchestration.runtime_config import deep_merge
 
-FIELDNAMES = [
-    "run",
-    "accelerator",
-    "model",
-    "version",
-    "prompt toks",
-    "output toks",
-    "TP",
-    "DP",
-    "EP",
-    "replicas",
-    "prefill_pod_count",
-    "decode_pod_count",
-    "router_config",
-    "measured concurrency",
-    "intended concurrency",
-    "measured rps",
-    "output_tok/sec",
-    "total_tok/sec",
-    "prompt_token_count_mean",
-    "prompt_token_count_p99",
-    "output_token_count_mean",
-    "output_token_count_p99",
-    "ttft_median",
-    "ttft_p95",
-    "ttft_p1",
-    "ttft_p999",
-    "tpot_median",
-    "tpot_p95",
-    "tpot_p99",
-    "tpot_p999",
-    "tpot_p1",
-    "itl_median",
-    "itl_p95",
-    "itl_p999",
-    "itl_p1",
-    "request_latency_median",
-    "request_latency_min",
-    "request_latency_max",
-    "successful_requests",
-    "errored_requests",
-    "uuid",
-    "ttft_mean",
-    "ttft_p99",
-    "itl_mean",
-    "itl_p99",
-    "runtime_args",
-    "guidellm_start_time_ms",
-    "guidellm_end_time_ms",
-    "image_tag",
-    "guidellm_version",
-    "mlflow_run_id",
-    "mlflow_experiment_id",
-    "notes",
-]
+# Analysis configuration for KPI regression testing
+analysis_config = AnalysisConfig(
+    comparison_labels=["product_version"],
+    ignored_labels=["cluster"],
+    sorting_labels=["product_version"],
+    regression_config={
+        "SCALAR_RELATIVE_CHANGE": {
+            "max_relative_regression": 0.15,  # 15% threshold (LLM performance can vary more)
+            "min_baseline_points": 2,  # Require at least 2 baseline points for reliability
+        },
+        "TWO_DIM_AUC_CHANGE": {
+            "max_relative_regression": 0.35,  # 35% threshold (LLM performance can vary more)
+            "min_baseline_points": 2,  # Require at least 2 baseline points for reliability
+        },
+    },
+)
 
 
 class LlmDGuideLLMPlugin(GuideLLMPlugin):
@@ -119,50 +78,10 @@ class LlmDGuideLLMPlugin(GuideLLMPlugin):
         output_path: Path,
         include_header_comments: bool = True,
     ) -> str:
-        def metadata_row(labels: dict[str, Any]) -> dict[str, Any]:
-            accelerator = labels.get("gpu_type") or labels.get("accelerator", "")
-            model = labels.get("hf_model_id") or labels.get("model_name", "")
-            run_model = model.replace("/", "-")
-            tp = labels.get("tensor_parallel_size") or labels.get("TP", "")
-            replicas = labels.get("replicas", "")
-            version = normalize_product_version(
-                labels.get("product_version") or labels.get("version", "")
-            )
-            deployment_profile = labels.get("deployment_profile", "")
-            if version and deployment_profile:
-                version = f"{version}-{deployment_profile}"
-            return {
-                "run": "-".join(str(value) for value in (accelerator, run_model, tp) if value),
-                "accelerator": accelerator,
-                "model": model,
-                "version": version,
-                "prompt toks": labels.get("prompt_toks", ""),
-                "output toks": labels.get("output_toks", ""),
-                "TP": tp,
-                "DP": labels.get("DP") or labels.get("data_parallel_size") or 0,
-                "EP": labels.get("EP") or labels.get("expert_parallel_size") or 0,
-                "replicas": replicas,
-                "prefill_pod_count": labels.get("prefill_pod_count", 0),
-                "decode_pod_count": labels.get("decode_pod_count", 0),
-                "router_config": labels.get("router_config", ""),
-                "uuid": labels.get("run_uuid", ""),
-                "runtime_args": labels.get("runtime_args", ""),
-                "guidellm_start_time_ms": labels.get("guidellm_start_time_ms", ""),
-                "guidellm_end_time_ms": labels.get("guidellm_end_time_ms", ""),
-                "image_tag": labels.get("image_tag", ""),
-                "guidellm_version": labels.get("guidellm_version", ""),
-                "mlflow_run_id": labels.get("mlflow_run_id", ""),
-                "mlflow_experiment_id": labels.get("mlflow_experiment_id", ""),
-                "notes": labels.get("notes", ""),
-            }
+        """Export KPI records to CSV format with llm-d dashboard schema."""
+        from . import csv_dashboard
 
-        return export_dashboard_kpis_to_csv(
-            kpi_records,
-            output_path,
-            prefix="llmd",
-            fieldnames=FIELDNAMES,
-            metadata_row=metadata_row,
-        )
+        return csv_dashboard.export_kpis_to_csv(kpi_records, output_path, include_header_comments)
 
 
 def get_plugin() -> PostProcessingPlugin:
